fix: allow underscore and forbid dashes in the databasename
